+++ src/game/Object/Player.cpp
@@ -4592,281 +4592,8 @@ void Player::InitDisplayIds()
     }
 }
 
-void Player::TakeExtendedCost(uint32 extendedCostId, uint32 count)
-{
-    ItemExtendedCostEntry const* extendedCost = sItemExtendedCostStore.LookupEntry(extendedCostId);
 
-    if (extendedCost->reqhonorpoints)
-    {
-        ModifyHonorPoints(-int32(extendedCost->reqhonorpoints * count));
-    }
-    if (extendedCost->reqarenapoints)
-    {
-        ModifyArenaPoints(-int32(extendedCost->reqarenapoints * count));
-    }
 
-    for (uint8 i = 0; i < MAX_EXTENDED_COST_ITEMS; ++i)
-    {
-        if (extendedCost->reqitem[i])
-        {
-            DestroyItemCount(extendedCost->reqitem[i], extendedCost->reqitemcount[i] * count, true);
-        }
-    }
-}
-
-// Return true is the bought item has a max count to force refresh of window by caller
-bool Player::BuyItemFromVendorSlot(ObjectGuid vendorGuid, uint32 vendorslot, uint32 item, uint8 count, uint8 bag, uint8 slot)
-{
-    // cheating attempt
-    if (count < 1)
-    {
-        count = 1;
-    }
-
-    // cheating attempt
-    if (bag != NULL_BAG && bag != INVENTORY_SLOT_BAG_0 && slot > MAX_BAG_SIZE && slot != NULL_SLOT)
-    {
-        return false;
-    }
-
-    if (!IsAlive())
-    {
-        return false;
-    }
-
-    ItemPrototype const* pProto = ObjectMgr::GetItemPrototype(item);
-    if (!pProto)
-    {
-        SendBuyError(BUY_ERR_CANT_FIND_ITEM, NULL, item, 0);
-        return false;
-    }
-
-    Creature* pCreature = GetNPCIfCanInteractWith(vendorGuid, UNIT_NPC_FLAG_VENDOR);
-    if (!pCreature)
-    {
-        DEBUG_LOG("WORLD: BuyItemFromVendor - %s not found or you can't interact with him.", vendorGuid.GetString().c_str());
-        SendBuyError(BUY_ERR_DISTANCE_TOO_FAR, NULL, item, 0);
-        return false;
-    }
-
-    VendorItemData const* vItems = pCreature->GetVendorItems();
-    VendorItemData const* tItems = pCreature->GetVendorTemplateItems();
-    if ((!vItems || vItems->Empty()) && (!tItems || tItems->Empty()))
-    {
-        SendBuyError(BUY_ERR_CANT_FIND_ITEM, pCreature, item, 0);
-        return false;
-    }
-
-    uint32 vCount = vItems ? vItems->GetItemCount() : 0;
-    uint32 tCount = tItems ? tItems->GetItemCount() : 0;
-
-    if (vendorslot >= vCount + tCount)
-    {
-        SendBuyError(BUY_ERR_CANT_FIND_ITEM, pCreature, item, 0);
-        return false;
-    }
-
-    VendorItem const* crItem = vendorslot < vCount ? vItems->GetItem(vendorslot) : tItems->GetItem(vendorslot - vCount);
-    if (!crItem)                                            // store diff item (cheating)
-    {
-        SendBuyError(BUY_ERR_CANT_FIND_ITEM, pCreature, item, 0);
-        return false;
-    }
-
-    if (crItem->item != item)                               // store diff item (cheating or special convert)
-    {
-        bool converted = false;
-
-        // possible item converted for BoA case
-        ItemPrototype const* crProto = ObjectMgr::GetItemPrototype(crItem->item);
-        if (crProto->Flags & ITEM_FLAG_BOA && crProto->RequiredReputationFaction &&
-                uint32(GetReputationRank(crProto->RequiredReputationFaction)) >= crProto->RequiredReputationRank)
-            converted = (sObjectMgr.GetItemConvert(crItem->item, getRaceMask()) != 0);
-
-        if (!converted)
-        {
-            SendBuyError(BUY_ERR_CANT_FIND_ITEM, pCreature, item, 0);
-            return false;
-        }
-    }
-
-    uint32 totalCount = pProto->BuyCount * count;
-
-    // check current item amount if it limited
-    if (crItem->maxcount != 0)
-    {
-        if (pCreature->GetVendorItemCurrentCount(crItem) < totalCount)
-        {
-            SendBuyError(BUY_ERR_ITEM_ALREADY_SOLD, pCreature, item, 0);
-            return false;
-        }
-    }
-
-    if (uint32(GetReputationRank(pProto->RequiredReputationFaction)) < pProto->RequiredReputationRank)
-    {
-        SendBuyError(BUY_ERR_REPUTATION_REQUIRE, pCreature, item, 0);
-        return false;
-    }
-
-    if (uint32 extendedCostId = crItem->ExtendedCost)
-    {
-        ItemExtendedCostEntry const* iece = sItemExtendedCostStore.LookupEntry(extendedCostId);
-        if (!iece)
-        {
-            sLog.outError("Item %u have wrong ExtendedCost field value %u", pProto->ItemId, extendedCostId);
-            return false;
-        }
-
-        // honor points price
-        if (GetHonorPoints() < (iece->reqhonorpoints * count))
-        {
-            SendEquipError(EQUIP_ERR_NOT_ENOUGH_HONOR_POINTS, NULL, NULL);
-            return false;
-        }
-
-        // arena points price
-        if (GetArenaPoints() < (iece->reqarenapoints * count))
-        {
-            SendEquipError(EQUIP_ERR_NOT_ENOUGH_ARENA_POINTS, NULL, NULL);
-            return false;
-        }
-
-        // item base price
-        for (uint8 i = 0; i < MAX_EXTENDED_COST_ITEMS; ++i)
-        {
-            if (iece->reqitem[i] && !HasItemCount(iece->reqitem[i], iece->reqitemcount[i] * count))
-            {
-                SendEquipError(EQUIP_ERR_VENDOR_MISSING_TURNINS, NULL, NULL);
-                return false;
-            }
-        }
-
-        // check for personal arena rating requirement
-        if (GetMaxPersonalArenaRatingRequirement(iece->reqarenaslot) < iece->reqpersonalarenarating)
-        {
-            // probably not the proper equip err
-            SendEquipError(EQUIP_ERR_CANT_EQUIP_RANK, NULL, NULL);
-            return false;
-        }
-    }
-
-    if (crItem->conditionId && !isGameMaster() && !sObjectMgr.IsPlayerMeetToCondition(crItem->conditionId, this, pCreature->GetMap(), pCreature, CONDITION_FROM_VENDOR))
-    {
-        SendBuyError(BUY_ERR_CANT_FIND_ITEM, pCreature, item, 0);
-        return false;
-    }
-
-    uint32 price = (crItem->ExtendedCost == 0 || pProto->Flags2 & ITEM_FLAG2_EXT_COST_REQUIRES_GOLD) ? pProto->BuyPrice * count : 0;
-
-    // reputation discount
-    if (price)
-    {
-        price = uint32(floor(price * GetReputationPriceDiscount(pCreature)));
-    }
-
-    if (GetMoney() < price)
-    {
-        SendBuyError(BUY_ERR_NOT_ENOUGHT_MONEY, pCreature, item, 0);
-        return false;
-    }
-
-    Item* pItem = NULL;
-
-    if ((bag == NULL_BAG && slot == NULL_SLOT) || IsInventoryPos(bag, slot))
-    {
-        ItemPosCountVec dest;
-        InventoryResult msg = CanStoreNewItem(bag, slot, dest, item, totalCount);
-        if (msg != EQUIP_ERR_OK)
-        {
-            SendEquipError(msg, NULL, NULL, item);
-            return false;
-        }
-
-        ModifyMoney(-int32(price));
-
-        if (crItem->ExtendedCost)
-        {
-            TakeExtendedCost(crItem->ExtendedCost, count);
-        }
-
-        pItem = StoreNewItem(dest, item, true);
-    }
-    else if (IsEquipmentPos(bag, slot))
-    {
-        if (totalCount != 1)
-        {
-            SendEquipError(EQUIP_ERR_ITEM_CANT_BE_EQUIPPED, NULL, NULL);
-            return false;
-        }
-
-        uint16 dest;
-        InventoryResult msg = CanEquipNewItem(slot, dest, item, false);
-        if (msg != EQUIP_ERR_OK)
-        {
-            SendEquipError(msg, NULL, NULL, item);
-            return false;
-        }
-
-        ModifyMoney(-int32(price));
-
-        if (crItem->ExtendedCost)
-        {
-            TakeExtendedCost(crItem->ExtendedCost, count);
-        }
-
-        pItem = EquipNewItem(dest, item, true);
-
-        if (pItem)
-        {
-            AutoUnequipOffhandIfNeed();
-        }
-    }
-    else
-    {
-        SendEquipError(EQUIP_ERR_ITEM_DOESNT_GO_TO_SLOT, NULL, NULL);
-        return false;
-    }
-
-    if (!pItem)
-    {
-        return false;
-    }
-
-    uint32 new_count = pCreature->UpdateVendorItemCurrentCount(crItem, totalCount);
-
-    WorldPacket data(SMSG_BUY_ITEM, 8 + 4 + 4 + 4);
-    data << pCreature->GetObjectGuid();
-    data << uint32(vendorslot + 1);                 // numbered from 1 at client
-    data << uint32(crItem->maxcount > 0 ? new_count : 0xFFFFFFFF);
-    data << uint32(count);
-    GetSession()->SendPacket(&data);
-
-    SendNewItem(pItem, totalCount, true, false, false);
-
-    return crItem->maxcount != 0;
-}
-
-uint32 Player::GetMaxPersonalArenaRatingRequirement(uint32 minarenaslot)
-{
-    // returns the maximal personal arena rating that can be used to purchase items requiring this condition
-    // the personal rating of the arena team must match the required limit as well
-    // so return max[in arenateams](min(personalrating[teamtype], teamrating[teamtype]))
-    uint32 max_personal_rating = 0;
-    for (int i = minarenaslot; i < MAX_ARENA_SLOT; ++i)
-    {
-        if (ArenaTeam* at = sObjectMgr.GetArenaTeamById(GetArenaTeamId(i)))
-        {
-            uint32 p_rating = GetArenaPersonalRating(i);
-            uint32 t_rating = at->GetRating();
-            p_rating = p_rating < t_rating ? p_rating : t_rating;
-            if (max_personal_rating < p_rating)
-            {
-                max_personal_rating = p_rating;
-            }
-        }
-    }
-    return max_personal_rating;
-}
 
 /**
  * @brief Updates the invalid-instance homebind timer and teleports when it expires.
